% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{myGrid}
\alias{myGrid}
\alias{print.myGrid}
\alias{summary.myGrid}
\alias{plot.myGrid}
\title{Methods for 'myGrid' objects.}
\usage{
\method{print}{myGrid}(x, ...)

\method{summary}{myGrid}(object, ...)

\method{plot}{myGrid}(x, neuron.col = "white", print.title = FALSE,
  the.titles = paste("Cluster", 1:prod(x$dim)), ...)
}
\arguments{
\item{x}{\code{myGrid} object}

\item{\dots}{Further arguments to the \code{\link{plot}} function.}

\item{object}{\code{myGrid} object}

\item{neuron.col}{Color(s) used to depict the neurons. Default value is 
\code{white}. If the argument is composed of one single color, neurons will 
all be filled with the same color. If the argument is composed of many 
colors, the number of colors must match the total number of neurons.}

\item{print.title}{Whether the cluster titles must be printed in center of
the grid or not. Default to \code{FALSE} (titles not displayed).}

\item{the.titles}{If \code{print.title = TRUE}, values of the title to 
display. Default to "Cluster " followed by the cluster number.}
}
\description{
Methods for the result of \code{\link{initGrid}}
(\code{myGrid} object)
}
\details{
The \code{myGrid} class has the following entries:
\itemize{
  \item{\code{coord}} 2-column matrix with x and y coordinates of the grid 
  units
  \item{\code{topo}} topology of the grid;
  \item{\code{dim}} dimensions of the grid (width corresponds to x 
  coordinates)
  \item{\code{dist.type}} distance type that defines the topology of the 
  grid.
}
During plotting, the color filling process uses the coordinates of the object
\code{x} included in \code{x$coord}.
}
\examples{
# creating grid
a.grid <- initGrid(dimension=c(5,5), topo="square", dist.type="maximum")

# plotting grid
# without any color specification
plot(a.grid)
# generating colors from rainbow() function
my.colors <- rainbow(5*5)
plot(a.grid, neuron.col=my.colors)
}
\seealso{
\code{\link{initGrid}} to define a \code{myGrid} class object.
}
\author{
Madalina Olteanu, \email{madalina.olteanu@univ-paris1.fr}\cr
Nathalie Vialaneix, \email{nathalie.vialaneix@inrae.fr}
}
