% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_som.R
\name{projectIGraph}
\alias{projectIGraph}
\alias{projectIGraph.somRes}
\title{Compute the projection of a graph on a grid}
\usage{
projectIGraph(object, init.graph, ...)
}
\arguments{
\item{object}{a \code{somRes} object.}

\item{init.graph}{an \link[igraph]{igraph} whose number of vertices is equal
to the clustering length of the \code{somRes} object.}

\item{\dots}{Not used.}
}
\value{
The result is an \code{\link[igraph]{igraph}} which vertexes are the
clusters (the clustering is thus understood as a vertex clustering) and the 
edges are the counts of edges in the original graph between two vertices
corresponding to the two clusters in the projected graph or, if 
\code{init.graph} is a weighted graph, the sum of the weights between the 
pairs of vertices corresponding to the two clusters.

The resulting igraph object's attributes are: \itemize{
  \item the graph attribute \code{layout} which provides the layout of the 
  projected graph according to the grid of the SOM;
  \item the vertex attributes \code{name} and \code{size} which, respectively
  are the vertex number on the grid and the number of vertexes included in 
  the corresponding cluster;
  \item the edge attribute \code{weight} which gives the number of edges (or 
  the sum of the weights) between the vertexes of the two corresponding 
  clusters.
}
}
\description{
Compute the projection of a graph, provided as an 
\code{\link[igraph]{igraph}} object, on the grid of the \code{somRes} object.
}
\examples{
data(lesmis)
set.seed(7383)
mis.som <- trainSOM(x.data=dissim.lesmis, type="relational", nb.save=10)
proj.lesmis <- projectIGraph(mis.som, lesmis)
\dontrun{plot(proj.lesmis)}
}
\references{
Olteanu M., Villa-Vialaneix N. (2015) Using SOMbrero for clustering and 
visualizing graphs. \emph{Journal de la Société Française de Statistique},
\strong{156}, 95-119.
}
\seealso{
\code{\link{projectIGraph.somSC}} which uses the results of a 
super-clustering to obtain another projected graph. \code{\link{plot.somRes}} 
with the option \code{type="graph"} or \code{\link{plot.somSC}} with the 
option \code{type="projgraph"}.
}
\author{
Madalina Olteanu \email{olteanu@ceremade.dauphine.fr}\cr
Nathalie Vialaneix \email{nathalie.vialaneix@inrae.fr}
}
