\name{sombreroGUI}
\alias{sombreroGUI}

\title{Graphical User Interface for SOMbrero}
\description{Starts the SOMbrero GUI}
\usage{
sombreroGUI()
}

\arguments{
  No arguments.
}

\value{This function starts the graphical user interface.
}

\author{Julien Boelaert \email{julien.boelaert@gmail.com}\cr
Nathalie Villa-Vialaneix \email{nathalie.villa@univ-paris1.fr}
}

\references{
shiny web page: http://www.rstudio.com/shiny/
}
