\name{sombreroGUI}
\alias{sombreroGUI}

\title{Graphical Web User Interface for SOMbrero}
\description{Starts the SOMbrero GUI}
\usage{
sombreroGUI()
}

\arguments{
  No arguments.
}

\value{This function starts the graphical user interface with the default 
system browser. This interface is more lickely to work properly with Firefox
\url{http://www.mozilla.org/fr/firefox/new/}. In case Firefox is not your 
default browser, copy/paste \url{http://localhost:8100/} into the address bar.
}

\author{Julien Boelaert \email{julien.boelaert@gmail.com}\cr
Madalina Olteanu \email{madalina.olteanu@univ-paris1.fr}\cr
Nathalie Villa-Vialaneix \email{nathalie.villa@toulouse.inra.fr}
}

\references{
Boelaert J., Bendhaiba L., Olteanu M., Villa-Vialaneix N. (2014) SOMbrero: an R
package for numeric and non-numeric self-organizing maps. In: T. Villmann, 
F.M. Schleif, M. Kaden & M. Lange (Eds.), \emph{Advances in Self-Organizing Maps
and Learning Vector Quantization} (Proceedings of WSOM 2014) (Vol. 295,
219-228). Mittweida, Germany: Berlin Springer Verlag. 

RStudio and Inc. (2013). shiny: Web Application Framework for R. R package
version 0.7.0. \url{http://cran.r-project.org/package=shiny}
}
