% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sombreroGUI.R
\name{sombreroGUI}
\alias{sombreroGUI}
\title{Graphical Web User Interface for SOMbrero}
\usage{
sombreroGUI()
}
\value{
This function starts the graphical user interface with the default 
system browser. This interface is more lickely to work properly with Firefox
\url{https://www.mozilla.org/fr/firefox/new/}. In case Firefox is not your 
default browser, copy/paste http://localhost:8100 into the URL bar.
}
\description{
Start the SOMbrero GUI.
}
\references{
Villa-Vialaneix N. (2017) Stochastic self-organizing map variants with the R
package SOMbrero. In: J.C. Lamirel, M. Cottrell, M. Olteanu, \emph{12th 
International Workshop on Self-Organizing Maps and Learning Vector 
Quantization, Clustering and Data Visualization (Proceedings of WSOM 2017)},
IEEE, Nancy, France.

RStudio and Inc. (2013). shiny: Web Application Framework for R. R package
version 0.7.0. \url{https://cran.r-project.org/package=shiny}
}
\author{
Élise Maigné <elise.maigne@inrae.fr>\cr
Julien Boelaert \email{julien.boelaert@gmail.com}\cr
Madalina Olteanu \email{olteanu@ceremade.dauphine.fr}\cr
Nathalie Vialaneix \email{nathalie.vialaneix@inrae.fr}
}
