\name{sombreroGUI}
\alias{sombreroGUI}

\title{Graphical Web User Interface for SOMbrero}
\description{Starts the SOMbrero GUI}
\usage{
sombreroGUI()
}

\arguments{
  No arguments.
}

\value{This function starts the graphical user interface with the default 
system browser. This interface is more lickely to work properly with Firefox
\url{http://www.mozilla.org/fr/firefox/new/}. In case Firefox is not your 
default browser, copy/paste \url{http://localhost:8100/} into the address bar.
}

\author{Julien Boelaert \email{julien.boelaert@gmail.com}\cr
Madalina Olteanu \email{madalina.olteanu@univ-paris1.fr}\cr
Nathalie Villa-Vialaneix \email{nathalie.villa@univ-paris1.fr}
}

\references{
RStudio and Inc. (2013). shiny: Web Application Framework for R. R package
version 0.7.0. \url{http://cran.r-project.org/package=shiny}
}
